added new header tabs for large screens
Large screens now have a list of the 4 big pages at the top, rather than the menu.
Nothing is changed for small screens.
This is done by having all the elements in the HTML files, but then
hiding them depending on the screen size.

One CSS rule was added to make the new tabs disappear before the page scrolls